% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{lsn_mcmc_estimate}
\alias{lsn_mcmc_estimate}
\title{MCMC Estimate Shotnoise Kernel}
\usage{
lsn_mcmc_estimate(
  phi,
  prior_lambda0 = c(1, 1000),
  prior_dbh = c(1, 3),
  prior_theta = NULL,
  prior_log_tau = c(0, 1),
  lambda0 = NULL,
  theta = NULL,
  log_tau = NULL,
  rho_scale = 1000,
  niter = 1000,
  print_every = -1,
  keep_every = 100,
  loc_storage = 10000,
  win = NULL,
  ada_bi = niter/2,
  prior_psi = c(gamma = 0, range = 0),
  Rmax = NULL,
  cutoff_q = 0.997,
  ...,
  timeout = Inf
)
}
\description{
MCMC Estimate Shotnoise Kernel
}
