% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root-kernel.R
\name{lsn_root_density_d_part}
\alias{lsn_root_density_d_part}
\title{Root Kernel D-part}
\usage{
lsn_root_density_d_part(d)
}
\arguments{
\item{d}{dbh, in meters}
}
\description{
The size-dependent part of the kernel.
}
