% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root-kernel.R
\name{lsn_root_density_r_part}
\alias{lsn_root_density_r_part}
\title{Root Kernel R-part}
\usage{
lsn_root_density_r_part(r, tau = 1)
}
\arguments{
\item{r}{distance argument}

\item{tau}{shape argument, >0}
}
\description{
The range-dependent part of the kernel.
}
