% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotnoise.R
\name{lsn_shotnoise}
\alias{lsn_shotnoise}
\title{Shotnoise Field Evaluator
Create a shotnoise field from given Psi pattern (x,y,dbh,tau)}
\usage{
lsn_shotnoise(psi, bbox, resx = 250, cutoff_q = 0.99, grid = NULL)
}
\description{
Shotnoise Field Evaluator
Create a shotnoise field from given Psi pattern (x,y,dbh,tau)
}
