% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_shotnoise_binary_marking.R
\name{lsn_simpars_default}
\alias{lsn_simpars_default}
\title{Default Simpars}
\usage{
lsn_simpars_default(
  bbox = rbind(0:1, 0:1) * 100,
  Rmax = 10,
  ndata = 100,
  tau = 1,
  cutoff_q = 0.99,
  psi_lambda0 = 0.01,
  psi_gamma = 0,
  psi_range = 5,
  psi_dbh = c(50, 200),
  theta = c(qlogis(0.15), log(2)),
  rho_scale = 1000,
  ...
)
}
\description{
Default Simpars
}
