% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_shotnoise_binary_marking.R
\name{lsn_simulate}
\alias{lsn_simulate}
\title{Simulate Latent Shotnoise Dependent Binary Marking}
\usage{
lsn_simulate(simpars = lsn_simpars_default(), xy, ...)
}
\arguments{
\item{simpars}{list of simulation parameters. See lsn_simpars_default}

\item{xy}{fixed locations of the trees to mark, ndata x 2 matrix of coordinates}

\item{...}{passed on to lsn_simpars_default}
}
\value{
list(phi=tibble of data points with marks, psi = latent generator points, simpars = simpars)
}
\description{
Simulate Latent Shotnoise Dependent Binary Marking
}
\details{
Todo
}
