% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_shotnoise_binary_marking.R
\name{lsn_simpars_default}
\alias{lsn_simpars_default}
\title{Default Simpars}
\usage{
lsn_simpars_default(
  bbox = rbind(0:1, 0:1) * 100,
  Rmax = 10,
  ndata = 100,
  tau = 1,
  cutoff_q = 0.99,
  psi_lambda0 = 0.01,
  psi_gamma = 0,
  psi_range = 5,
  psi_dbh = c(50, 200),
  theta = c(qlogis(0.15), log(2)),
  rho_scale = 1000,
  ...
)
}
\arguments{
\item{bbox}{bounding box for the simulation. Like rbind(xlim, ylim) (dimension on rows)}

\item{Rmax}{Buffer for extending bbox}

\item{ndata}{Number of points}

\item{tau}{Kernel parameter}

\item{cutoff_q}{Kernel parameter}

\item{psi_lambda0}{Generator intensity}

\item{psi_dbh}{Generator dbh parameters, vector(2), Gamma distribution}

\item{theta}{Coefficient of the logistic regression, (intercept, shotnoise)}

\item{rho_scale}{scaling of the field effect, default 1000.}

\item{...}{}
}
\description{
Default Simpars
}
